+) 增加仓库模块功能，完善商品模块功能

+++ onlineshop-goods/src/main/java/com/hlju/onlineshop/goods/dto/SkuDTO.java
@@ -20,7 +20,7 @@ public class SkuDTO {
     private String skuSubtitle;
     private List<ImageDTO> images;
     private List<String> descar;
-    private BigDecimal fullCount;
+    private Integer fullCount;
     private BigDecimal discount;
     private Integer countStatus;
     private BigDecimal fullPrice;
